% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{add_outer_labels}
\alias{add_outer_labels}
\title{Add outer labels to a multi-panel plot}
\usage{
add_outer_labels(P, main = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{P}{Parameter list}

\item{main}{Main title}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}
}
\description{
Add outer labels to a multi-panel plot
}
\keyword{internal}
