% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{apply_par}
\alias{apply_par}
\title{Apply graphical parameters that are valid for par()}
\usage{
apply_par(P)
}
\arguments{
\item{P}{Parameter list from get_plot_params()}
}
\value{
Previous par settings (invisibly)
}
\description{
Apply graphical parameters that are valid for par()
}
\keyword{internal}
