% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affectOU-methods.R
\name{coef.affectOU}
\alias{coef.affectOU}
\title{Extract model parameters}
\usage{
\method{coef}{affectOU}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{affectOU}}.}

\item{...}{Additional arguments (unused).}
}
\value{
A list containing the model parameters: \code{theta}, \code{mu}, \code{gamma}, and
\code{sigma}. For 1D models, these are returned as single numbers. For
multivariate models, they are returned as matrices.
}
\description{
Extract the parameters of an affect Ornstein-Uhlenbeck (OU) model as a list.
}
\examples{
model <- affectOU(ndim = 2)
coef(model)
}
\seealso{
Model parameters are explained in \code{\link[=affectOU]{affectOU()}} and can be modified using \code{\link[=update.affectOU]{update()}}.
}
\concept{config}
