% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R
\name{explore.affectOU}
\alias{explore.affectOU}
\title{Explore method for affectOU}
\usage{
\method{explore}{affectOU}(x, ...)
}
\arguments{
\item{x}{An \code{affectOU} object}

\item{...}{Additional arguments (unused)}
}
\description{
Explore method for affectOU
}
