% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{filter_args}
\alias{filter_args}
\title{Filter a list of arguments}
\usage{
filter_args(args, valid_args)
}
\arguments{
\item{args}{Named list of arguments}

\item{valid_args}{Character vector of valid argument names}
}
\value{
Filtered named list
}
\description{
Filter a list of arguments
}
\keyword{internal}
