% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{generate_shades}
\alias{generate_shades}
\title{Generate shades of a colour}
\usage{
generate_shades(color, n, min_alpha = 0.3, max_alpha = 1)
}
\arguments{
\item{color}{Base colour}

\item{n}{Number of shades}

\item{min_alpha}{Minimum alpha (transparency)}

\item{max_alpha}{Maximum alpha (transparency)}
}
\value{
Character vector of colours
}
\description{
Generate shades of a colour
}
\keyword{internal}
