% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{get_layout}
\alias{get_layout}
\title{Calculate layout dimensions for multi-panel plots}
\usage{
get_layout(n, P, user_args = list(), by_dim = TRUE, call = rlang::caller_env())
}
\arguments{
\item{n}{Number of panels needed}

\item{P}{Parameter list containing nrow, ncol, mfrow, or mfcol}

\item{user_args}{Original user arguments (to check for mfrow/mfcol)}

\item{by_dim}{Logical; plot each dimension in separate panel?}

\item{call}{Environment used to report errors. Relevant only when
calling this function from another function, so that errors name the
function the user called.}
}
\value{
List with nrow and ncol
}
\description{
Calculate layout dimensions for multi-panel plots
}
\keyword{internal}
