% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{get_lims}
\alias{get_lims}
\title{Calculate axis limits for multiple panels}
\usage{
get_lims(
  data,
  ndim,
  nsim,
  lim = NULL,
  share_axis = FALSE,
  include = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data array}

\item{ndim}{Number of dimensions}

\item{nsim}{Number of simulations}

\item{lim}{User-supplied lim (NULL, vector, or list)}

\item{share_axis}{Logical; share axis across panels?}
}
\value{
List of lim vectors
}
\description{
Calculate axis limits for multiple panels
}
\keyword{internal}
