% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{get_valid_args}
\alias{get_valid_args}
\title{Get valid argument names for a function}
\usage{
get_valid_args(fun, include_par = TRUE)
}
\arguments{
\item{fun}{Function or function name}

\item{include_par}{Logical; include graphical parameters from par()?}
}
\value{
Character vector of valid argument names
}
\description{
Get valid argument names for a function
}
\keyword{internal}
