% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{logLik.fit_affectOU}
\alias{logLik.fit_affectOU}
\title{Extract log-likelihood from fitted OU affect model}
\usage{
\method{logLik}{fit_affectOU}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link[=fit.affectOU]{fit_affectOU}}.}

\item{...}{Additional arguments (unused).}
}
\value{
An object of class \code{logLik} with \code{df} and \code{nobs} attributes.
}
\description{
Extract log-likelihood from fitted OU affect model
}
\examples{
model <- affectOU(theta = 0.5, mu = 0, gamma = 1)
sim <- simulate(model, stop = 500, dt = 0.01, save_at = 0.1)
data <- as.data.frame(sim)
fitted <- fit(model, data = data$value, times = data$time)
logLik(fitted)
}
\concept{fit}
