% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{ou_plot_fit_acf}
\alias{ou_plot_fit_acf}
\title{Plot ACF of residuals}
\usage{
ou_plot_fit_acf(
  fit,
  lag.max = 30,
  palette = "Dark 3",
  main = "ACF of Residuals",
  xlab = "Lag",
  ylab = "ACF",
  alpha = 0.8,
  ...
)
}
\arguments{
\item{fit}{A \code{fit_affectOU} object}

\item{lag.max}{Maximum lag to compute. Specified in terms of saved time units. For example, \code{lag.max = 10} with \code{save_at = 0.1} corresponds to a lag of 10 time units and 10/0.1=100 time points.}

\item{palette}{Color palette. Should be one of \code{\link[grDevices:hcl.pals]{grDevices::hcl.pals()}}.}

\item{main}{Main title}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{alpha}{Alpha transparency for colors (0 = transparent, 1 = opaque)}

\item{...}{Additional graphical parameters}
}
\value{
NULL (invisibly), called for side effects only.
}
\description{
Assess remaining autocorrelation in the residuals of the fitted model by
plotting the autocorrelation function (ACF). Ideally, there should be no
significant autocorrelation remaining, indicating that the model has
adequately captured the temporal dependencies in the data.
}
\examples{
model <- affectOU(theta = 0.5, mu = 0, gamma = 1)
sim <- simulate(model, stop = 500, dt = 0.01, save_at = 0.1)
data <- as.data.frame(sim)
fitted <- fit(model, data = data$value, times = data$time)
ou_plot_fit_acf(fitted)
}
\concept{fit}
