% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{print.simulate_affectOU}
\alias{print.simulate_affectOU}
\title{Print simulation object}
\usage{
\method{print}{simulate_affectOU}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{simulate_affectOU} object}

\item{digits}{Number of digits for numeric display}

\item{...}{Additional arguments (unused)}
}
\description{
Display overview of an \code{simulate_affectOU} object without printing
the full data array.
}
