% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affectOU-methods.R
\name{print.summary_affectOU}
\alias{print.summary_affectOU}
\title{Print summary of affectOU model}
\usage{
\method{print}{summary_affectOU}(x, digits = 3, max_dim = 20, ...)
}
\arguments{
\item{x}{An object of class \code{summary_affectOU}}

\item{digits}{Number of digits to display}

\item{max_dim}{Maximum number of dimensions to display details for}

\item{...}{Additional arguments (unused)}
}
\description{
Print summary of affectOU model
}
\examples{
model <- affectOU(ndim = 2)
summary_model <- summary(model)
print(summary_model)
}
