% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{setup_panel}
\alias{setup_panel}
\title{Set up axes for a plot panel}
\usage{
setup_panel(xlim, ylim, P, main = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{xlim}{X-axis limits}

\item{ylim}{Y-axis limits}

\item{P}{Parameter list}

\item{main}{Main title (NULL to suppress)}

\item{xlab}{X-axis label (NULL to suppress)}

\item{ylab}{Y-axis label (NULL to suppress)}
}
\value{
None (invisible)
}
\description{
Set up axes for a plot panel
}
\keyword{internal}
