% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpaper.R
\docType{data}
\name{simpaper}
\alias{simpaper}
\title{Simulation in Evers & Vanhasbroeck (2026)}
\format{
An object of class \code{\link[=simulate.affectOU]{simulate_affectOU}} containing simulated trajectories from the specified Ornstein-Uhlenbeck process.
}
\usage{
simpaper
}
\description{
An uncoupled two-dimensional Ornstein-Uhlenbeck process simulated with \code{affectOU}.
}
\details{
The process is defined by the stochastic differential equation:
\deqn{d\mathbf{X}_t = \mathbf{\Theta}(\mathbf{\mu} - \mathbf{X}_t)dt + \mathbf{\Gamma}d\mathbf{W}_t}

where:
\itemize{
\item \eqn{\mathbf{\Theta} = \begin{pmatrix} 1 & 0 \\ 0 & 1 \end{pmatrix}} is the drift matrix
\item \eqn{\mathbf{\mu} = \begin{pmatrix} 0 \\ 0 \end{pmatrix}} is the equilibrium mean vector
\item \eqn{\mathbf{\Gamma} = \begin{pmatrix} 1 & 0 \\ 0 & 1 \end{pmatrix}} is the diffusion matrix
\item \eqn{\mathbf{W}_t} is a two-dimensional Wiener process
}
}
\concept{simulate}
\keyword{datasets}
