% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affectOU-methods.R
\name{dim.affectOU}
\alias{dim.affectOU}
\title{Extract dimensionality}
\usage{
\method{dim}{affectOU}(x)
}
\arguments{
\item{x}{An object of class \code{affectOU}}
}
\value{
Integer, the dimensionality of the process.
}
\description{
Extract the number of dimensions of an affect Ornstein-Uhlenbeck (OU) model.
}
\examples{
model <- affectOU()
dim(model)
}
