% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{ou_plot_fit_qq}
\alias{ou_plot_fit_qq}
\title{QQ plot of residuals}
\usage{
ou_plot_fit_qq(
  fit,
  palette = "Temps",
  main = "Normal Q-Q Plot of Residuals",
  xlab = "Theoretical Quantiles",
  ylab = "Sample Quantiles",
  alpha = 0.8,
  ...
)
}
\arguments{
\item{fit}{A \code{fit_affectOU} object}

\item{palette}{Color palette. Should be one \code{\link[grDevices:palettes]{grDevices::hcl.pals()}}.}

\item{main}{Main title}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{alpha}{Alpha transparency for colors (0 = transparent, 1 = opaque)}

\item{...}{Additional graphical parameters}
}
\value{
NULL (invisibly), called for side effects only.
}
\description{
Visualize a normal Q-Q plot of the residuals to assess their distributional properties. Ideally, residuals should lie approximately along the reference line, indicating normality.
}
\examples{
model <- affectOU(theta = 0.5, mu = 0, gamma = 1)
sim <- simulate(model, stop = 500, dt = 0.01, save_at = 0.1)
data <- as.data.frame(sim)
fitted <- fit(model, data = data$value, times = data$time)
ou_plot_fit_qq(fitted)
}
