% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulate.R
\name{plot_parameters}
\alias{plot_parameters}
\title{Plot parameter documentation}
\arguments{
\item{which_dim}{Dimension indices to plot (NULL for all)}

\item{which_sim}{Simulation indices to plot (NULL for all)}

\item{by_dim}{Logical; plot each dimension in separate panel?}

\item{palette}{Color palette. Should be one of \code{\link[grDevices:hcl.pals]{grDevices::hcl.pals()}}.}

\item{alpha}{Alpha transparency for colors (0 = transparent, 1 = opaque)}

\item{lwd}{Line width}

\item{share_xaxis}{Logical; use same x-axis limits for all panels?}

\item{share_yaxis}{Logical; use same y-axis limits for all panels?}

\item{freq}{Logical; plot frequency instead of density?}

\item{breaks}{Number of histogram breaks}

\item{lag.max}{Maximum lag to compute. Specified in terms of saved time units. For example, \code{lag.max = 10} with \code{save_at = 0.1} corresponds to a lag of 10 time units and 10/0.1=100 time points.}

\item{main}{Main title}

\item{sub}{Subtitle for panels}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{legend_position}{Position of legend (one of \code{"bottomright"}, \code{"bottom"},
\code{"bottomleft"}, \code{"left"}, \code{"topleft"}, \code{"top"}, \code{"topright"}, \code{"right"},
\code{"center"}, \code{"none"}). Set to \code{"none"} to hide legend.}

\item{...}{Additional graphical parameters}
}
\description{
Plot parameter documentation
}
\keyword{internal}
