% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affectOU-methods.R
\name{print.affectOU}
\alias{print.affectOU}
\title{Print affect OU model}
\usage{
\method{print}{affectOU}(x, digits = 3, max_dim = 20, ...)
}
\arguments{
\item{x}{An object of class \code{affectOU}}

\item{digits}{Number of digits to display}

\item{max_dim}{Maximum number of dimensions to display details for}

\item{...}{Additional arguments (unused)}
}
\description{
Print affect OU model
}
\examples{
model <- affectOU(ndim = 2)
print(model)
}
