% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R, R/fit.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{equation}
\alias{fit}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{equation}}, \code{\link[generics]{fit}}}
}}

